refactor(skill): split registry persistence owner
