+++ framework/core/src/python/kungfu/skill/_registry/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""Private responsibility owners for the public Skill registry."""
